requirements.txt
bottle~=0.12
sqlalchemy==1.4.2
jsonpickle==2.0.0
requests==2.25.1
psycopg2==2.8.6
pgdb==0.0.8
python-dotenv==0.17.0
